Tests/LibWeb/Text/expected/HTML/document-close-readystatechange-replaces-parser.txt

Replaced parser during interactive readystatechange event
Ready state after close: loading
Ready state in first task: loading
Ready state in second task: loading
Replacement iframe load event fired after close
